refactor: split in folders and rename files
