+++ lib/ui/views/levels/blood_test_chart_page/bar_chart_graph.dart
@@ -2,7 +2,7 @@ import 'dart:math' as math;
 
 import 'package:fl_chart/fl_chart.dart';
 import 'package:flutter/material.dart';
-import 'package:mona/ui/views/chart/bar_groups.dart';
+import 'package:mona/ui/views/levels/bar_groups.dart';
 
 class BarChartGraph extends StatelessWidget {
   final List<FlSpot> spots;
