+++ lib/ui/views/levels/level_entry_spots.dart
