+++ lib/ui/views/levels/levels_page/levels_page.dart
@@ -13,11 +13,11 @@ import 'package:mona/i18n/helpers/units_l10n.dart';
 import 'package:mona/i18n/translations.g.dart';
 import 'package:mona/services/preferences_service.dart';
 import 'package:mona/ui/constants/dimensions.dart';
-import 'package:mona/ui/views/chart/baby_graph.dart';
-import 'package:mona/ui/views/chart/baby_main_graph.dart';
-import 'package:mona/ui/views/chart/blood_tests_chart_page.dart';
-import 'package:mona/ui/views/chart/chart_page.dart';
-import 'package:mona/ui/views/chart/level_entry_spots.dart';
+import 'package:mona/ui/views/levels/levels_page/baby_bar_chart_graph.dart';
+import 'package:mona/ui/views/levels/levels_page/baby_main_chart_graph.dart';
+import 'package:mona/ui/views/levels/blood_test_chart_page/blood_tests_chart_page.dart';
+import 'package:mona/ui/views/levels/main_graph_page/chart_page.dart';
+import 'package:mona/ui/views/levels/level_entry_spots.dart';
 import 'package:mona/ui/widgets/main_page_wrapper.dart';
 import 'package:mona/util/time_difference.dart';
 import 'package:provider/provider.dart';
