+++ lib/ui/views/levels/main_graph_page/chart_page.dart
@@ -1,8 +1,8 @@
 import 'package:flutter/material.dart';
 import 'package:mona/data/providers/medication_intake_provider.dart';
 import 'package:mona/i18n/translations.g.dart';
-import 'package:mona/ui/views/chart/chart_graph.dart';
-import 'package:mona/ui/views/chart/chart_slider.dart';
+import 'package:mona/ui/views/levels/main_graph_page/chart_graph.dart';
+import 'package:mona/ui/views/levels/main_graph_page/chart_slider.dart';
 import 'package:mona/ui/widgets/main_page_wrapper.dart';
 import 'package:provider/provider.dart';
