+++ lib/ui/views/levels/main_graph_page/chart_slider.dart
