+++ lib/ui/views/main_tabs.dart
@@ -2,8 +2,8 @@
 import 'package:flutter/material.dart';
 import 'package:material_symbols_icons/material_symbols_icons.dart';
 import 'package:mona/i18n/translations.g.dart';
-import 'package:mona/ui/views/chart/blood_test_page.dart';
-import 'package:mona/ui/views/chart/levels_page.dart';
+import 'package:mona/ui/views/levels/blood_tests_page/blood_test_page.dart';
+import 'package:mona/ui/views/levels/levels_page/levels_page.dart';
 import 'home/home_page.dart';
 import 'home/settings/settings_page.dart';
 import 'intakes/choose_schedule_page.dart';
